entity-bookings/entity-payments

+++ src/repository/entity/court_number.ts
@@ -2,16 +2,16 @@ import { Entity, PrimaryGeneratedColumn, Column, OneToMany } from 'typeorm';
 
 @Entity({ name: 'court' })
 export class CourtNumber {
-  @PrimaryGeneratedColumn({ name: 'court_id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'court_id', type: 'int8' })
   court_id!: number;
 
-  @Column({ name: 'stadium_id', type: 'integer' })
+  @Column({ name: 'stadium_id', type: 'int4' })
   stadium_id!: number;
 
-  @Column({ name: 'court_number', type: 'integer' })
+  @Column({ name: 'court_number', type: 'int4' })
   court_number!: number;
 
-  @Column({ name: 'price_hour', type: 'integer' })
+  @Column({ name: 'price_hour', type: 'int4' })
   price_hour!: number;
 
   @Column({ name: 'active', type: 'varchar', length: 50 })
